make pattern blocks bigger
